requirements.txt
ollama
groq
streamlit
pandas
python-dotenv
pyperclip  # for copy-to-clipboard functionality